Use a context to respond with the following event if the user wants it
Achieved by setting a context on the first response, including the date
of the event that we mentioned last, as well as a talk index since
php[world] has multiple talks in the same slot. While there are events
left in the schedule, we keep passing revised context values to allow
for iterating through the event list.